move client fixture to conftest
