+++ tests/test_magstats.py
@@ -1,6 +1,3 @@
-from fixtures import client
-
-
 def test_magstats(psql_service, client):
     r = client.get("objects/ZTF1/magstats")
     assert r.status_code == 200
